# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/balance": {
            "get": {
                "description": "Obtém o saldo disponível do usuário autenticado",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Retorna o saldo do usuário",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token JWT",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Saldo do usuário",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "number"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/buy": {
            "post": {
                "description": "Permite ao usuário comprar ações, com validações de saldo e tipo de ordem",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Compra ações",
                "parameters": [
                    {
                        "description": "Dados da ordem de compra",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.BuyOrder"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes da compra",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Insufficient funds",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Market is closed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/generate-token": {
            "get": {
                "description": "Retorna um token JWT válido para um usuário específico",
                "summary": "Gera um token JWT para um usuário",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nome do usuário",
                        "name": "username",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Indica se o usuário é admin",
                        "name": "isAdmin",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Token JWT gerado",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erro ao gerar token",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "Autentica um usuário e retorna um token JWT",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Realiza login do usuário",
                "parameters": [
                    {
                        "description": "Credenciais do usuário",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.Credentials"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Token JWT",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Could not create token",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/market-stocks": {
            "get": {
                "description": "Obtém uma lista de ações do mercado americano com seus preços de mercado",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Retorna a lista de ativos do mercado americano e seus preços",
                "responses": {
                    "200": {
                        "description": "Lista de ações e preços",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api.MarketStock"
                            }
                        }
                    }
                }
            }
        },
        "/order-status/{order_id}": {
            "get": {
                "description": "Retorna o status atual de uma ordem de compra pelo ID da ordem",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Consulta o status de uma ordem",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID da ordem",
                        "name": "order_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status da ordem",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Ordem não encontrada",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/register": {
            "post": {
                "description": "Registra um novo usuário na plataforma",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Cria um novo usuário",
                "parameters": [
                    {
                        "description": "Dados do usuário",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User registered",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Could not hash password",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/users": {
            "get": {
                "description": "Retorna a lista de usuários cadastrados com suas senhas (apenas administradores podem acessar)",
                "produces": [
                    "application/json"
                ],
                "summary": "Lista usuários cadastrados",
                "responses": {
                    "200": {
                        "description": "Lista de usuários cadastrados",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Forbidden: Only admins can access this endpoint",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/validate-token": {
            "get": {
                "description": "Verifica se um token JWT é válido",
                "summary": "Valida um token JWT",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token JWT",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Claims do token",
                        "schema": {
                            "$ref": "#/definitions/api.TokenClaims"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.BuyOrder": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "order_type": {
                    "type": "string"
                },
                "stock": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "api.Credentials": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "api.MarketStock": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "symbol": {
                    "type": "string"
                }
            }
        },
        "api.TokenClaims": {
            "type": "object",
            "properties": {
                "admin": {
                    "type": "boolean",
                    "example": true
                },
                "exp": {
                    "type": "integer"
                },
                "username": {
                    "type": "string",
                    "example": "test_user"
                }
            }
        },
        "api.User": {
            "type": "object",
            "properties": {
                "admin": {
                    "type": "boolean"
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    }
}